added day 24
i copied the solution from a python dev bro. i am ashamed
but wtf who would look at the input file. my initial solution
had 200 lines and then i realised that the input range was of
the order of 10 triliards or whatever. and my solution would take
literally 10 million years to finish, so i opened up reddit and
got spoiled by the fact that the input was a cycle, and out of
all the instructions only 2 of them mattered for this problem
the fourth one and the fifteenth one, then i saw a solution that
took 10 lines of code in python and copied it, i still don't understand
why a stack solves this problem. help. send help. i suck at coding